requirements/base.txt
django==1.11.9
djangorestframework==3.5.1
MySQL-python==1.2.5
gunicorn==19.7.1
django-cors-headers==2.1.0
requests
